--- packages/commonwealth/client/scripts/views/pages/view_proposal/types.ts
@@ -1,9 +1,0 @@
-export type Prefetch = {
-  [identifier: string]: {
-    commentsStarted: boolean;
-    pollsStarted: boolean;
-    viewCountStarted: boolean;
-    profilesStarted: boolean;
-    profilesFinished: boolean;
-  };
-};
