+++ packages/commonwealth/client/styles/components/content_comments_tree.scss
@@ -1,6 +1,6 @@
 @import 'shared';
 
-.ProposalComments {
+.ContentCommentsTree {
   display: flex;
   flex-direction: column;
 }
