+++ packages/commonwealth/client/styles/pages/view_proposal/proposal_body.scss
@@ -8,44 +8,6 @@
   max-width: 768px;
   width: 100%;
 
-  .user-avatar {
-    margin: -6px 6px -6px 0;
-  }
-
-  @include mediumSmallInclusive {
-    margin: 0 40px 0 0;
-  }
-
-  .header {
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-
-    .info-and-menu-row {
-      align-items: center;
-      display: flex;
-      flex-wrap: wrap;
-      gap: 8px;
-
-      .username {
-        font-weight: 500;
-
-        @include caption;
-      }
-
-      .header-text.Text {
-        color: $neutral-500;
-      }
-    }
-  }
-
-  .proposal-links {
-    align-items: center;
-    display: flex;
-    flex-wrap: wrap;
-    gap: 8px;
-  }
-
   .proposal-content {
     display: flex;
     flex-direction: column;
@@ -56,28 +18,6 @@
       background-color: $neutral-200;
       border-radius: $border-radius-corners;
       padding: 16px;
-    }
-
-    .proposal-response-row {
-      width: 100%;
-      display: flex;
-      max-width: 300px;
-      align-items: center;
-      margin-top: 16px;
-
-      > * {
-        margin-right: 8px;
-      }
-    }
-  }
-
-  .onchain-status-text.Text {
-    &.pass {
-      color: $green-600;
-    }
-
-    &.fail {
-      color: $rorange-600;
     }
   }
